feat: add attribute passing for angular and stencil
