+++ packages/core/src/__tests__/data/ref/basic-forwardRef-metadata.raw.tsx
