+++ packages/core/src/__tests__/data/ref/basic-ref-assignment.raw.tsx
