+++ packages/core/src/__tests__/data/ref/basic-ref-attribute-passing.raw.tsx
@@ -1,0 +1,13 @@
+import { useMetadata, useRef } from '@builder.io/mitosis';
+
+useMetadata({
+  attributePassing: {
+    enabled: true,
+  },
+});
+
+export default function BasicRefAttributePassingComponent() {
+  const buttonRef = useRef<HTMLButtonElement | null>(null);
+
+  return <button ref={buttonRef}>Attribute Passing</button>;
+}
